refactor(shared-types): utilities/ の helper を責務ベースの subdir 構成に再編（SPR-69）
4 つの grab-bag な *-helpers.ts を、責務を反映した subdir 構成（12 ファイル / 7 subdir）に再編:

- discord-helpers.ts → discord/avatar.ts, discord/guild-membership.ts,
  user/display-name.ts, user/role-label.ts, formatters/relative-time.ts
- creator-helpers.ts → creator/type-label.ts, validators/dlsite-ids.ts
- evaluation-helpers.ts → evaluation/characteristic-axes.ts, evaluation/aggregator.ts
- search-filter-helpers.ts → search/filter-state.ts, search/filter-descriptors.ts,
  search/date-range-preset.ts

公開 API (index.ts の re-export) は不変。全 import は型のみ (import type) で
ランタイム Zod 依存は無し。

検証:
- typecheck (monorepo 全体) / lint / test 全 pass
  (shared-types 913, ui 332, functions 340, web 688)
- web build pass、全 prerender ページの First Load JS から Zod-lib chunk 排除を維持
  (/works/[workId] のみ pre-existing。本番と chunk hash 一致で SPR-69 起因ではない)
