+++ packages/shared-types/src/utilities/search/filter-descriptors.ts
@@ -1,74 +1,4 @@
-import type { DateRangePreset, UnifiedSearchFilters } from "./search-filters";
-
-export function getDateRangeFromPreset(preset: DateRangePreset): {
-	from: Date;
-	to: Date;
-} {
-	const now = new Date();
-	const today = new Date(now.getFullYear(), now.getMonth(), now.getDate());
-
-	switch (preset) {
-		case "today":
-			return {
-				from: today,
-				to: new Date(today.getTime() + 24 * 60 * 60 * 1000),
-			};
-
-		case "this_week": {
-			const dayOfWeek = today.getDay();
-			const daysToMonday = dayOfWeek === 0 ? 6 : dayOfWeek - 1;
-			const weekStart = new Date(today.getTime() - daysToMonday * 24 * 60 * 60 * 1000);
-			return {
-				from: weekStart,
-				to: new Date(weekStart.getTime() + 7 * 24 * 60 * 60 * 1000),
-			};
-		}
-
-		case "this_month": {
-			const monthStart = new Date(now.getFullYear(), now.getMonth(), 1);
-			const monthEnd = new Date(now.getFullYear(), now.getMonth() + 1, 1);
-			return {
-				from: monthStart,
-				to: monthEnd,
-			};
-		}
-
-		case "last_30_days": {
-			const thirtyDaysAgo = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-			return {
-				from: thirtyDaysAgo,
-				to: now,
-			};
-		}
-
-		case "custom":
-			return {
-				from: today,
-				to: now,
-			};
-	}
-}
-
-export function hasActiveFilters(filters: UnifiedSearchFilters): boolean {
-	return (
-		!!filters.dateRange ||
-		!!filters.dateFrom ||
-		!!filters.dateTo ||
-		filters.playCountMin !== undefined ||
-		filters.playCountMax !== undefined ||
-		filters.likeCountMin !== undefined ||
-		filters.likeCountMax !== undefined ||
-		filters.favoriteCountMin !== undefined ||
-		filters.favoriteCountMax !== undefined ||
-		filters.durationMin !== undefined ||
-		filters.durationMax !== undefined ||
-		(filters.tags?.length ?? 0) > 0 ||
-		(filters.playlistTags?.length ?? 0) > 0 ||
-		(filters.userTags?.length ?? 0) > 0 ||
-		(filters.categoryNames?.length ?? 0) > 0 ||
-		filters.sortBy !== "relevance"
-	);
-}
+import type { DateRangePreset, UnifiedSearchFilters } from "../search-filters";
 
 function getDateRangeDescription(filters: UnifiedSearchFilters): string | null {
 	if (filters.dateRange && filters.dateRange !== "custom") {
