+++ packages/shared-types/src/utilities/user/display-name.ts
@@ -1,0 +1,7 @@
+export function resolveDisplayName(
+	displayName: string | undefined,
+	globalName: string | undefined,
+	username: string,
+): string {
+	return displayName || globalName || username;
+}
