+++ packages/shared-types/src/utilities/user/role-label.ts
@@ -1,0 +1,8 @@
+export function getUserRoleLabel(role: "member" | "moderator" | "admin"): string {
+	const labels = {
+		member: "メンバー",
+		moderator: "モデレーター",
+		admin: "管理者",
+	};
+	return labels[role];
+}
