+++ packages/shared-types/src/utilities/validators/dlsite-ids.ts
@@ -1,0 +1,7 @@
+export function isValidCircleId(circleId: string): boolean {
+	return /^RG\d+$/.test(circleId);
+}
+
+export function isValidCreatorId(creatorId: string): boolean {
+	return creatorId.length > 0;
+}
